requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
